Fixed GTS Features + Bug Fixes
Changed the DNS server to use pkmnclassic.net's DNS server. This allows for the NDS to be able to once again connect to GTS communication.

Also fixed a handfull of broken features in the sendpkm.py file that either read the input path wrong, or simply had a typo and read the wrong variable.